Move base classes for testing to embulk-junit4 from embulk-core:tests
