+++ embulk-filter-remove_columns/build.gradle
@@ -29,7 +29,6 @@ dependencies {
     testImplementation project(":embulk-api")
     testImplementation project(":embulk-spi")
     testImplementation project(":embulk-core")
-    testImplementation project(":embulk-core").sourceSets.test.output
     testImplementation project(":embulk-deps")
     testImplementation project(":embulk-junit4")
     testImplementation project(":embulk-formatter-csv")
