+++ embulk-formatter-csv/build.gradle
@@ -35,7 +35,6 @@ dependencies {
     testImplementation project(":embulk-api")
     testImplementation project(":embulk-spi")
     testImplementation project(":embulk-core")
-    testImplementation project(":embulk-core").sourceSets.test.output
     testImplementation project(":embulk-deps")
     testImplementation project(":embulk-junit4")
 }
