+++ embulk-guess-csv/build.gradle
@@ -31,7 +31,6 @@ dependencies {
     testImplementation project(":embulk-api")
     testImplementation project(":embulk-spi")
     testImplementation project(":embulk-core")
-    testImplementation project(":embulk-core").sourceSets.test.output
     testImplementation project(":embulk-deps")
     testImplementation project(":embulk-junit4")
