+++ embulk-junit4/src/main/java/org/embulk/EmbulkTestRuntime.java
