+++ embulk-junit4/src/main/java/org/embulk/spi/MockParserPlugin.java
